Refactor nft Valuation view
